--- common/src/main/kotlin/observable/net/NetConsts.kt
@@ -1,7 +1,0 @@
-package observable.net
-
-import net.minecraft.resources.ResourceLocation
-
-object NetConsts {
-    val INIT_TPS_PROFILER = ResourceLocation("id/observable/initTps")
-}
