+++ common/src/main/kotlin/observable/util/Constants.kt
@@ -5,7 +5,7 @@ import net.minecraft.network.chat.ClickEvent
 import net.minecraft.network.chat.Component
 
 const val MOD_URL = "https://o.tas.sh/"
-val MOD_URL_COMPONENT = Component.literal(MOD_URL)
+val MOD_URL_COMPONENT: Component = Component.literal(MOD_URL)
     .withStyle(ChatFormatting.UNDERLINE)
     .withStyle {
         it.withClickEvent(ClickEvent(ClickEvent.Action.OPEN_URL, MOD_URL))
